storage_service: boost shares for streaming write during bootstrap, replace
During bootstrap and replace operations the node can't take reads and
we'd like to see the process ending ASAP. This is because until the
process ends, we keep having to duplicate writes to an extended set. Not
to mention, in the case of a cluster expansion users want to use the
added capacity sooner rather than later.

Streaming generates a lot of compaction activity, that competes with the
bootstrap itself, slowing it down.

Long term, we are moving to treat those compactions differently and
maybe postpone them altogether. However in the meantime we can play with
the shares so as to allow streaming writes to proceed with equal
footing.

Streaming has the lowest shares of all, which makes sense when it is a
maintenance operation. But in the case of a bootstrap or replace
streaming is the star of the show.

Signed-off-by: Glauber Costa <glauber@scylladb.com>