cdc: wait for all seeds' tokens when joining the ring without bootstrap
It's a common scenario in tests where a new cluster is created by starting
multiple nodes as seeds concurrently, even though the recommended way to
start nodes in production is sequential.

When a node joins a Scylla cluster for the first time (i.e. it joins the
token ring), it proposes a new CDC generation which is supposed to
include tokens of all nodes: all nodes that were already present plus
the new node. The new node learns the tokens of other nodes by
contacting a seed node (other than itself, if it's also a seed).

However, in the concurrent startup scenario, the above strategy doesn't really
work. If there are 3 nodes joining the ring concurrently, then the
second and third might not see each other's tokens when proposing their
CDC generations, so we're left with some generation which doesn't
include all tokens.

Fortunately, (most of) these tests use a certain pattern of specifying
seed lists:
- node 1 specifies {1}
- node 2 specifies {1, 2}
- node 3 specifies {1, 2, 3}
etc.
Exploiting that fact, we can prevent the race by forcing node N to wait
until it sees the tokens of nodes {1, ..., N - 1} (which are all seeds).

This commit does that. Obviously, this is a heuristic, but it's all we
can have before implementing a real solution, which is transactional
joining of nodes.