cdc: when proposing a new generation, make sure it's the latest one
When a new node joins the token ring it proposes a new generation of CDC
streams. The new generation uses the tokens of all nodes currently in
the ring. When nodes learn about the new generation they'll start using
it because it's the latest one (it has the greatest timestamp).

However, it's possible that the previous generation was proposed very
recently (due to nodes starting concurrently) and due to desynchronized
clocks, the latest generation will actually have a smaller timestamp
than the previous one, so nodes will just ignore it and keep using the
previous one.

This commit minimalizes the chances of this happening. It's in the
spirit of heuristically preventing bad things happening before we
implement a real, correct solution (transactional joining of nodes).