configure: Warn on large stacks
This adds a warning with a different limit in each mode. The limit is
picked as 1KiB lower than the value where no warning would be print.

This makes it easy to spot the worse offender. With that we can either
fix it or silence the warning once we are sure we can handle large
frames in that context.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200311205300.324383-1-espindola@scylladb.com>