range_streamer: Fix nr_ranges_per_stream_plan calculation
It is supposed to stream 10% of the ranges per stream plan and finishes
the whole ranges in 10 stream plans. However, it will end up with stream
11 stream plans.

For example, we have 769 ranges in total.

nr_ranges_per_stream_plan = 769 / 10 = 76

10 stream plan would only stream 76 * 10 = 760 ranges

As a result, we need another stream plan to stream the rest of the 769
- 760 = 9 ranges.

To fix, we round up

nr_ranges_per_stream_plan = (769 + 9) / 10  = 77

so that, 10 stream plan would stream all of them.

Refs: #5109