cql3: when altering table, keep old values of unchanged extensions
When the user performed

alter ks.t with compaction = {...}

the values of most other options, which were not specified in the
statement, e.g. compression, were left unchanged. That wasn't true for
extension options however: for example, the "cdc" option was removed.

This commit fixes the behavior to keep the old values of extension
options not specified in the alter statement.