Make compaction cancellation easier.
Right now we can cancel all compactions of a specific type (cleanup,
regular, etc). This is not good enough, as in some situations we may
want to cancel a specific compaction.

One example is the upcoming offstrategy compactions that are fired
when the compaction strategy changes. If the user changes the compaction
strategy and then changes it again we will want to cancel just the
compaction that we started. Not all compactions on the column family,
not all compactions of a specific type.

There is already a mechanism to cancel a compaction given a task
pointer, so all we have to do is expose the task pointer to the
caller.

At this moment I don't see the need to cancel regular minor compactions
so that is deferred until and if we need that.

For now we will return the task to major compaction callers.

Note that there is no need to do anything with repect to the management
of the _tasks list. If the task is not added to the list the future will
be ready immediately. If it is, the compaction will have started already
for sure and will stop with an exception - removing itself from the
list, when it sees the stop notification

Signed-off-by: Glauber Costa <glauber@scylladb.com>