cdc+counters: make counters work in CDC
This patch augments counter "deltas" on leader node. Before that,
"deltas" are "sanitized", which means that they are set to 0 if
they try to update a deleted counter, so CDC will not contain
spurious (ineffective) updates.

On the CDC side, counter columns are represented as `long_type`,
with unchanged logic behind pre-/postimage (preimage query returns
ordinary integer, postimage is calculated from delta and preimg).